Add new entity type: TreasuryAsset
This new entity will track the individual performance and values of each asset held by the treasury. Token balance, RFV, Market Value are all calculated and stored within this entity. The sum total for the treasury is updated as a part of generating/updating the asset details.

One driving factor this this addition is the fact that we track multiple values for the same asset, not just dollar value. We are interested in the total tonnes of carbon held by the treasury for example.