Update graph-cli and graph-ts version. Add bonds and Klimate balance updates.
While troubleshooting an issue with building the forked Klimate balance code from Olympus, it was discovered that breaking changes in v0.22.0 of graph-ts was the cause. Followed the steps outlined here (https://thegraph.com/docs/en/developer/assemblyscript-migration-guide) to migrate the subgraph to the new version of AssemblyScript.

Bond Datasources are added for all three currently active Klima bonds.

Klimate balances update to include these bonding transactions. Staking transactions are also added in this commit.

Bond discounts are now calculated as part of the ProtocolMetrics update.

As all of the originally needed data sources are included in this commit, the subgraph-orig.yaml file is removed.

Add Bonded Discount to the Deposit entity

This adds the discount at time of bonding to the deposit entity. Could be useful for tracking a klimate's own performance and overall bond discounts over time