Break out bond data to its own subgraph
Define the bond specific subgraph. This includes an updated schema that only triggers
on bonding transactions, and only holds bonder totals. Useful to aid in sync time and
overall indexing performance on hosted services.

Add BCV and second MCO2 deployments

Add the `handleBCV` function to track the BCVs on a daily basis for bonds.

Create a new datasource for the re-deployed MCO2 bonds due to debt falling to zero